+++ Midterm_k-means_CUDA/csvHandler.h
@@ -4,6 +4,6 @@ using namespace std;
 
 float* readCsv();
 
-void writeCsv(float* points, float* centroids, float* clusters, int iteration, int k);
+void writeCsv(float* points, float* centroids, int* clusters, int iteration, int k);
 
 void initialize();
